Simplify documentation about Pandoc operation
If something needs to be this in-depth and specific it should probably
be documented in the implementation to stop it going out-of-date.

Just summarise the Lua parts instead, the target rules are fairly
clear now that the GNU Make specific functions are gone, detailed
explanations on the Lua code should be coming from the Lua files.