Topic builder for web, new desktop layout, general improvements
The topic builder relies on a md file providing the hierarchy, currently index.md, but maybe we could repurpose _Sidebar.md.

There's a left sidebar with the topics, and a right sidebar with the page TOC.
When there is no TOC, or an empty one, the article is centered keeping a max width similar to the original.

Topic navigation offers expandable sections. I'm undecided whether the expand arrows should be right or left

I haven't worked on mobile yet. I will probably move some stuff inside an hamburger menu.